tests/hil/scripts/pytest-hil/pyproject.toml
[build-system]
requires = ["hatchling>=1.19"]
build-backend = "hatchling.build"

[project]
name = "pytest-hil"
version = "0.1.0"
dependencies = [
  'allure-pytest',
  'esptool',
  'pylink-square',
  'pynrfjprog',
  'pyserial',
  'PyYAML',
  'trio-serial',
]
classifiers = [
    "Framework :: Pytest",
]

[project.entry-points.pytest11]
hil = "pytest_hil.plugin"
